requirements.txt
python-telegram-bot
requests
python-dotenv
beautifulsoup4
lxml
aiohttp
aiofiles
yt-dlp
